requirements.txt
streamlit==1.29.0
Pillow==10.1.0
pandas==2.0.3
